feat: 수익률 계산 기능 구현 및 예외 처리 후 재입력 기능 구현
수익률 계산하여 결과 출력에 이를 포함시켰으며,
잘못된 값 입력 시에, 예외 발생 후 다시 그 지점부터 재입력하도록 로직을 변경했습니다.

+++ src/main/java/lotto/domain/Prize.java
@@ -1,21 +1,21 @@
 package lotto.domain;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public enum Prize {
-    THREE(BigInteger.valueOf(5_000)),
-    FOUR(BigInteger.valueOf(50_000)),
-    FIVE(BigInteger.valueOf(1_500_000)),
-    FIVE_BONUS(BigInteger.valueOf(30_000_000)),
-    SIX(BigInteger.valueOf(2_000_000_000));
+    THREE(BigDecimal.valueOf(5_000)),
+    FOUR(BigDecimal.valueOf(50_000)),
+    FIVE(BigDecimal.valueOf(1_500_000)),
+    FIVE_BONUS(BigDecimal.valueOf(30_000_000)),
+    SIX(BigDecimal.valueOf(2_000_000_000));
 
-    private final BigInteger prizeMoney;
+    private final BigDecimal prizeMoney;
 
-    Prize(BigInteger prizeMoney) {
+    Prize(BigDecimal prizeMoney) {
         this.prizeMoney = prizeMoney;
     }
 
-    public BigInteger getPrizeMoney() {
+    public BigDecimal getPrizeMoney() {
         return prizeMoney;
     }
 }
